dev > main: fixed transformation and plotting of special points
