+++ src/models/internal/points.py
@@ -12,6 +12,9 @@ from ...thirdparty.types import *
 from ...core.poly import *
 from ..generated.internal import *
 
+# NOTE: foreign import
+from ..generated.app import SpecialPointsConfig
+
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 # EXPORTS
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -22,6 +25,7 @@ __all__ = [
     'get_renormalised_polynomial_time_only',
     'get_renormalised_polynomial_values_only',
     'get_renormalised_data',
+    'get_renormalised_coordinates_of_special_points',
 ]
 
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -141,3 +145,23 @@ def get_renormalised_data(
         {'time[orig]': 'float', 'time': 'float', quantity: 'float'}
     )
     return data
+
+
+def get_renormalised_coordinates_of_special_points(
+    points: dict[str, SpecialPointsConfig],
+    p: list[float],
+    info: FittedInfo,
+) -> list[tuple[str, SpecialPointsConfig]]:
+    '''
+    NOTE: assumes polynomial has been renormalised.
+    '''
+    T = info.normalisation.period
+    points_ = [(key, point.copy()) for key, point in points.items()]
+
+    times = T * np.asarray([point.time for _, point in points_])
+    values = poly(times, *p)
+    for t, y, (_, point) in zip(times, values, points_):
+        point.time = t
+        point.value = y
+
+    return points_
